Split up `ActivationConfig`
