+++ src/Whim.CommandPalette/Config/CommandPaletteMenuActivationConfig.cs
@@ -1,0 +1,12 @@
+namespace Whim.CommandPalette;
+
+/// <summary>
+/// Config for activating the command palette with a menu.
+/// </summary>
+public record CommandPaletteMenuActivationConfig : BaseCommandPaletteActivationConfig
+{
+	/// <summary>
+	/// The matcher to use to filter the results.
+	/// </summary>
+	public ICommandPaletteMatcher Matcher { get; init; } = new MostRecentlyUsedMatcher();
+}
